requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-lru==1.0.2
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
discord.py==1.7.2
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
google==3.0.0
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
lyricsgenius==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
mysql-connector-python==8.0.25
oauthlib==3.1.0
packaging==20.9
progress==1.5
progressbar==2.5
protobuf==3.15.8
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
PyMySQL==1.0.2
pyparsing==2.4.7
Pyrogram==1.2.9
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.20
tidalapi==0.6.8
tidaloauth4mopidy==0.2.0
tmdbv3api==1.7.5
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.7
webencodings==0.5.1
Werkzeug==2.0.1
yarl==1.6.3
